move common delete logic into shared code
